+++ presentation/src/main/kotlin/com/ddd/oi/presentation/recommendedlist/RecommendedListNavGraph.kt
@@ -10,7 +10,7 @@ fun NavController.navigateToRecommendedList() {
 }
 
 fun NavGraphBuilder.recommendedListNavGraph(
-    navigateToRecommendedDetail: () -> Unit
+    navigateToRecommendedDetail: (Long) -> Unit
 ) {
     composable<Route.RecommendedList> {
         RecommendedListScreen(
